requirements.txt
asgiref==3.3.1
Django==3.1.7
django-ckeditor==6.0.0
django-js-asset==1.2.2
django-mptt==0.11.0
django-taggit==1.3.0
Pillow==8.1.0
python-decouple==3.4
pytz==2021.1
sorl-thumbnail==12.7.0
sqlparse==0.4.1
